// src/interfaces/appIterfaces.tsx
// Generated by https://quicktype.io

export interface LoginData {
    correo: string,
    contraseña: string;
}

export interface RegisterData {
    nombre: string,
    correo: string,
    contraseña: string;
    usuario: string
}

export interface RegisterResponse{
    mensaje:string
}
export interface LoginResponse {
    mensaje:string,
    usuario: Usuario;
    token:   string;
    rol:string
}

export interface Usuario {
    _id:        string;
    nombre:     string;
    id:         string;
    correo:     string;
    contraseña: string;
    usuario:    string;
    __v:        number;
}
// Generated by https://quicktype.io

export interface Profile {
    status:   string;
    profiles: ProfileElement[];
}

export interface ProfileElement {
    _id:       string;
    date:      string;
    nombre:    string;
    apellido:  string;
    email:     string;
    telefono:  string;
    direccion: string;
    image:     string;
    __v:       number;
}

// Generated by https://quicktype.io

export interface RespProfile {
    status:  string;
    message: string;
}


// Generated by https://quicktype.io

export interface Profileresul {
    status:  string;
    product: Product;
}

export interface Product {
    _id:       string;
    date:      string;
    nombre:    string;
    apellido:  string;
    email:     string;
    telefono:  string;
    direccion: string;
    estatus:   string;
    image:     null;
    __v:       number;
}
export interface ProducttResponseAll {
    status:   string;
    products: Products[];
  }
  
  export interface Products {
    _id:          string;
    date:         string;
    nombre:       string;
    ingredientes: string;
    sabores:      string;
    categoria:    string;
    datos:        string;
    peso:         string;
    capacidad:    string;
    precio:       string;
    image:        null | string;
    __v:          number;
  }
// Generated by https://quicktype.io

export interface Pedidos {
    status: string;
    sales:  Sale[];
}

export interface Sale {
    _id:             string;
    date:            string;
    solicitante:     Solicitante;
    producto_nombre: string;
    cantidad:        string;
    telefono:        string;
    direccion:       Direccion;
    datos:           Datos;
    estado:          Estado;
    total:           string;
    image:           Estado | null;
    __v:             number;
}

export enum Datos {
    EnProcesoDePago = "En proceso de pago",
    Pagado = "Pagado",
}

export enum Direccion {
    ProlongacionPatoni5 = "Prolongacion patoni 5",
}

export enum Estado {
    Empaquetado = "Empaquetado",
    Entregado = "Entregado",
    Null = "null",
}

export enum Solicitante {
    LuisDavidCasanovaChavez = "Luis David Casanova Chavez",
}